Move vara_manager to research tools
