+++ varats/varats/tools/driver_build_setup.py
@@ -15,12 +15,12 @@ from varats.tools.research_tools.research_tool import (
     ResearchTool,
     SpecificCodeBase,
 )
+from varats.tools.research_tools.vara_manager import BuildType
 from varats.utils.cli_util import (
     get_research_tool,
     get_supported_research_tool_names,
     initialize_cli_tool,
 )
-from varats.vara_manager import BuildType
 
 
 class VaRATSSetup:
