+++ varats/varats/tools/research_tools/research_tool.py
@@ -5,9 +5,7 @@ import abc
 import typing as tp
 from pathlib import Path
 
-from varats.utils.filesystem_util import FolderAlreadyPresentError
-from varats.utils.logger_util import log_without_linesep
-from varats.vara_manager import (
+from varats.tools.research_tools.vara_manager import (
     BuildType,
     add_remote,
     branch_has_upstream,
@@ -25,6 +23,8 @@ from varats.vara_manager import (
     show_status,
     update_all_submodules,
 )
+from varats.utils.filesystem_util import FolderAlreadyPresentError
+from varats.utils.logger_util import log_without_linesep
 
 
 class SubProject():
