+++ varats/varats/tools/research_tools/vara.py
@@ -17,14 +17,14 @@ from varats.tools.research_tools.research_tool import (
     ResearchTool,
     SubProject,
 )
-from varats.utils.exceptions import ProcessTerminatedError
-from varats.utils.logger_util import log_without_linesep
-from varats.vara_manager import (
+from varats.tools.research_tools.vara_manager import (
     BuildType,
     ProcessManager,
     run_process_with_output,
     set_vara_cmake_variables,
 )
+from varats.utils.exceptions import ProcessTerminatedError
+from varats.utils.logger_util import log_without_linesep
 
 LOG = logging.getLogger(__name__)
